add user authorization to /todos route

+++ src/services/todos.service.ts
@@ -4,15 +4,16 @@ interface Todo {
     id: string,
     state: string,
     description: string,
-    createdat: Date,
-    completedat: Date | null
+    created_at: Date,
+    completed_at: Date | null,
+    user_id: string
 }
 
-export const getTodos = async (filter: 'ALL' | 'COMPLETE' | 'INCOMPLETE', orderBy: 'DESCRIPTION' | 'CREATED_AT' | 'COMPLETED_AT') => {
+export const getTodos = async (userId: string, filter: 'ALL' | 'COMPLETE' | 'INCOMPLETE', orderBy: 'DESCRIPTION' | 'CREATED_AT' | 'COMPLETED_AT') => {
 
     return await new Promise<Todo[]>((resolve, reject) => {
 
-        const queryBuilder = pg('todos');
+        const queryBuilder = pg('todos').where({ user_id: userId });
 
         if (filter === 'COMPLETE' || filter === 'INCOMPLETE') {
             queryBuilder.where({ state: filter });
@@ -20,8 +21,8 @@ export const getTodos = async (filter: 'ALL' | 'COMPLETE' | 'INCOMPLETE', orderB
 
         switch (orderBy) {
             case 'DESCRIPTION': queryBuilder.orderBy('description', 'asc'); break;
-            case 'COMPLETED_AT': queryBuilder.orderBy('completedat', 'asc'); break;
-            default: queryBuilder.orderBy('createdat', 'asc'); break;
+            case 'COMPLETED_AT': queryBuilder.orderBy('completed_at', 'asc'); break;
+            default: queryBuilder.orderBy('created_at', 'asc'); break;
         }
 
         queryBuilder.then((data) => resolve(data as Todo[]))
@@ -29,16 +30,16 @@ export const getTodos = async (filter: 'ALL' | 'COMPLETE' | 'INCOMPLETE', orderB
     });
 };
 
-export const getTodo = async (id: string) => {
+export const getTodo = async (userId: string, id: string) => {
 
     return await new Promise<Todo | null>((resolve, reject) => {
 
-        const queryBuilder = pg('todos').where({ id }).returning([
+        const queryBuilder = pg('todos').where({ id, user_id: userId }).returning([
             'id',
             'state',
             'description',
-            'createdat',
-            'completedat'
+            'created_at',
+            'completed_at'
         ]).first();
 
         queryBuilder.then((todo) => {
@@ -49,61 +50,66 @@ export const getTodo = async (id: string) => {
     });
 };
 
-export const postTodo = async (description: string) => {
+export const postTodo = async (userId: string, description: string) => {
 
     return await new Promise<Todo>((resolve, reject) => {
 
         const queryBuilder = pg('todos')
-            .insert({ description })
+            .insert({ description, user_id: userId })
             .returning([
                 'id',
                 'state',
                 'description',
-                'createdat',
-                'completedat'
+                'created_at',
+                'completed_at',
+                'user_id'
             ]);
 
-        queryBuilder.then((todos) => resolve(todos[0]))
+        queryBuilder
+            .then((todos) => resolve(todos[0]))
             .catch((e) => reject(e));
     });
 };
 
 type patchContent = { state?: 'COMPLETE' | 'INCOMPLETE', description?: string }
-export const patchTodo = async (id: string, content: patchContent) => {
+export const patchTodo = async (userId: string, id: string, content: patchContent) => {
 
     // TODO: completedAt could be updated via a pg trigger on 'state' change
-    const completedat = content.state === 'COMPLETE' ? new Date() : null;
+    const completed_at = content.state === 'COMPLETE' ? new Date() : null;
 
     return await new Promise<Todo>((resolve, reject) => {
 
         const queryBuilder = pg('todos')
-            .where({ id })
-            .update({ ...content, completedat }, [
+            .where({ id, user_id: userId })
+            .update({ ...content, completed_at }, [
                 'id',
                 'state',
                 'description',
-                'createdat',
-                'completedat'
+                'created_at',
+                'completed_at'
             ]);
 
-        queryBuilder.then((todos) => resolve(todos[0] as Todo))
+        queryBuilder
+            .then((todos) => resolve(todos[0] as Todo))
             .catch((e) => reject(e));
     });
 };
 
-export const deleteTodo = async (id: string) => {
+export const deleteTodo = async (userId: string, id: string) => {
 
     return await new Promise<number>((resolve, reject) => {
 
         const queryBuilder = pg('todos');
 
-        queryBuilder.delete().where({ id });
+        queryBuilder.delete().where({ id, user_id: userId });
 
-        queryBuilder.then((deleteCount) => {
+        queryBuilder
+            .then((deleteCount) => {
 
-            if (typeof deleteCount === 'number') {
-                resolve(deleteCount);
-            }
-        }).catch((e) => reject(e));
+                if (typeof deleteCount === 'number') {
+                    resolve(deleteCount);
+                }
+            })
+            .catch((e) => reject(e));
     });
 };
